Use dequeueReusableCellWithClass(_:) in TokenListViewController
